requirements.txt
tensorflow-gpu
sklearn
pandas
numpy
tables
liac-arff